Improve enabling, disabling, and showing status for blocklists. (#2)
* Improve enabling, disabling, and showing status for blocklists.

Add chain_exists function to check if an iptables chain already exists.

When enabling UFW, check if each blocklist chain already exists and for each one that does already exist, remove it and then recreate the chain. Prevents issues starting UFW if a chain already existed when you tried to enable UFW.

When disabling UFW, check to make sure each chain actually exists before removing them. Otherwise, if a chain was already removed, `ufw disable` will error and stop cleaning up the other blocklists.

Reformat the status display so that it separates each chain and shows the column headers for what each value is.

* Add a comment to each DROP

Added a comment to each DROP to make the status a little easier to read.