Adjust maxelem to suit the list. Fixes #10
Set the maximum number of elements allowed in the ipset dynamically based on the number of entries in the blocklist. A configurable variable is provided to increase the maximum number of elements to give some headroom to allow the admin to manually add some entries. The default is 128. This variable must be set in after.init and the cron job ufw-blocklist-ipsum